feasted on PostCreationFragmentTest, cannont test capture_button
